// src/types/public/Clinics.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

import { type UsersId } from './Users';
import { ClinicTreatment, ClinicType, type ClinicTypesId } from './ClinicTypes';
import { type DistrictsId } from './Districts';

/** Identifier type for public.clinics */
export type ClinicsId = number & { __brand: 'ClinicsId' };

/** Represents the table public.clinics */
export default interface Clinics {
  id: ClinicsId;

  name: string;

  spoken_languages: [];

  description: string | null;

  representative_id: UsersId | null;

  confirmed_at: Date | null;

  type_id: ClinicTypesId | null;

  payment_info: unknown | null;
  district_id: DistrictsId | null;
  address: string | null;
  lat: string | null;
  lng: string | null;
  social_links: unknown | null;
  keywords: unknown | null;
  seo_settings: unknown | null;
  email: string | null;
  phone: string | null;
  contact_person: string | null;
  contact_person_phone: string | null;
  url: string;
  dtsanalpos_id: number | null;
  treatment_price: ClinicTreatment;
  created_at: Date;
  clinic_type: ClinicType;
  updated_at: Date;
  review_count: number;
  review_rating: string;
  cities_name: string | null
  districts_name: string | null
  cover_image: string | null
  treatment_category_name: string | null
  max_fee: number | null
  min_fee: number | null
}

/** Represents the initializer for the table public.clinics */
export interface ClinicsInitializer {
  /** Default value: nextval('clinics_id_seq'::regclass) */
  id?: ClinicsId;

  name: string;

  /** Default value: '["tr"]'::jsonb */
  spoken_languages?: unknown | null;

  description?: string | null;

  representative_id?: UsersId | null;

  confirmed_at?: Date | null;

  type_id?: ClinicTypesId | null;

  payment_info?: unknown | null;

  district_id?: DistrictsId | null;

  address?: string | null;

  lat?: string | null;

  lng?: string | null;

  /** Default value: '{}'::jsonb */
  social_links?: unknown | null;

  /** Default value: '[]'::jsonb */
  keywords?: unknown | null;

  /** Default value: '{}'::jsonb */
  seo_settings?: unknown | null;

  email?: string | null;

  phone?: string | null;

  contact_person?: string | null;

  contact_person_phone?: string | null;

  url?: string | null;

  dtsanalpos_id?: number | null;

  /** Default value: CURRENT_TIMESTAMP */
  created_at?: Date;

  /** Default value: CURRENT_TIMESTAMP */
  updated_at?: Date;
}

/** Represents the mutator for the table public.clinics */
export interface ClinicsMutator {
  id?: ClinicsId;

  name?: string;

  spoken_languages?: unknown | null;

  description?: string | null;

  representative_id?: UsersId | null;

  confirmed_at?: Date | null;

  type_id?: ClinicTypesId | null;

  payment_info?: unknown | null;

  district_id?: DistrictsId | null;

  address?: string | null;

  lat?: string | null;

  lng?: string | null;

  social_links?: unknown | null;

  keywords?: unknown | null;

  seo_settings?: unknown | null;

  email?: string | null;

  phone?: string | null;

  contact_person?: string | null;

  contact_person_phone?: string | null;

  url?: string | null;

  dtsanalpos_id?: number | null;

  created_at?: Date;

  updated_at?: Date;
}